chore: Move brand endpoint to endpoints section of app
