+++ webapp/admin/views.py
@@ -497,34 +497,6 @@ def delete_policy(store_id: str, model_name: str, revision: str):
     return make_response(res, 500)
 
 
-@admin.route("/api/store/<store_id>/brand")
-@login_required
-@exchange_required
-def get_brand_store(store_id: str):
-    brand_id = get_brand_id(flask.session, store_id)
-    res = {}
-    try:
-        brand = publisher_gateway.get_brand(flask.session, brand_id)
-
-        res["data"] = brand
-        res["success"] = True
-
-    except StoreApiResponseErrorList as error_list:
-        res["success"] = False
-        res["message"] = " ".join(
-            [
-                f"{error.get('message', 'An error occurred')}"
-                for error in error_list.errors
-            ]
-        )
-        res["data"] = []
-
-    response = make_response(res)
-    response.cache_control.max_age = 3600
-
-    return response
-
-
 @admin.route("/api/store/<store_id>/signing-keys")
 @login_required
 @exchange_required
